implemented camera

+++ Engine/source/Engine/Graphics/GPUResourceManager.h
@@ -12,7 +12,7 @@ namespace Okay
 	typedef uint16_t ResourceHandle;
 
 	constexpr uint32_t INVALID_AH = INVALID_UINT32;
-	constexpr uint32_t INVALID_RH = INVALID_UINT16;
+	constexpr uint16_t INVALID_RH = INVALID_UINT16;
 
 	struct ResourceAllocation
 	{
